refactor: manage utils for CmsEditor
